refactor(mdx-code): wrap `Pre` and `Code` components
issue #776
