+++ components/MDX/MDXCode.tsx
@@ -1,0 +1,7 @@
+import { Code } from '@components/CodeBlocks';
+
+interface Props {}
+
+const MDXCode = (props: Props) => <Code {...props} />;
+
+export default MDXCode;
