+++ components/MDX/MDXPre.tsx
@@ -1,0 +1,7 @@
+import { Pre } from '@components/CodeBlocks';
+
+interface Props {}
+
+const MDXPre = (props: Props) => <Pre {...props} />;
+
+export default MDXPre;
